+++ is-chrome.js
@@ -1,6 +1,9 @@
 'use strict';
 
+var isEdge = require('./is-edge');
+var isSamsung = require('./is-samsung');
+
 module.exports = function isChrome(ua) {
   ua = ua || navigator.userAgent;
-  return ua.indexOf('Chrome') !== -1 || ua.indexOf('CriOS') !== -1;
+  return (ua.indexOf('Chrome') !== -1 || ua.indexOf('CriOS') !== -1) && !isEdge(ua) && !isSamsung(ua);
 };