+++ is-mobile-firefox.js
@@ -1,0 +1,8 @@
+'use strict';
+
+var isIosFirefox = require('./is-ios-firefox');
+
+module.exports = function isMobileFirefox(ua) {
+  ua = ua || global.navigator.userAgent;
+  return isIosFirefox(ua) || /iPhone|iPod|iPad|Mobile|Tablet/i.test(ua) && /Firefox/i.test(ua);
+};