+++ test/supports-popups.js
@@ -22,7 +22,11 @@ describe('supportsPopups', function () {
     var key, ua;
     var keysToSkip = [
       'androidOperaMini',
-      'iPhoneUnsupportedChrome'
+      'iPhoneUnsupportedChrome',
+      'androidSamsungUnsupported',
+      'iPadFirefox',
+      'iPodFirefox',
+      'iPhoneFirefox'
     ];
 
     for (key in AGENTS) {
@@ -53,5 +57,9 @@ describe('supportsPopups', function () {
 
   it('returns true for iOS Chrome greather than 48', function () {
     expect(supportsPopups(AGENTS.iPhoneSupportedChrome)).to.equal(true);
+  });
+
+  it('returns false for old unsupported Samsung browser', function () {
+    expect(supportsPopups(AGENTS.androidSamsungUnsupported)).to.equal(false);
   });
 });